+++ plugins/sv-inspector/src/SvInspectorView/components/SvInspectorView.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { observer } from 'mobx-react'
-import { Grid, FormControlLabel, Checkbox } from '@mui/material'
 import { makeStyles } from 'tss-react/mui'
 import { ResizeHandle } from '@jbrowse/core/ui'
 
 // locals
 import { SvInspectorViewModel } from '../models/SvInspectorView'
+import CircularViewOptions from './CircularViewOptions'
 
 const useStyles = makeStyles()(theme => ({
   resizeHandleVert: {
@@ -26,46 +26,10 @@ const useStyles = makeStyles()(theme => ({
   viewsContainer: {
     display: 'flex',
   },
-  spreadsheetViewContainer: {
-    borderRight: '1px solid #ccc',
+  container: {
     overflow: 'hidden',
   },
-  circularViewOptions: {
-    padding: theme.spacing(1),
-  },
 }))
-
-const CircularViewOptions = observer(function ({
-  svInspector,
-}: {
-  svInspector: SvInspectorViewModel
-}) {
-  const { classes } = useStyles()
-
-  return (
-    <Grid
-      container
-      className={classes.circularViewOptions}
-      style={{ height: svInspector.circularViewOptionsBarHeight }}
-    >
-      <Grid item>
-        <FormControlLabel
-          control={
-            <Checkbox
-              checked={svInspector.onlyDisplayRelevantRegionsInCircularView}
-              onChange={e =>
-                svInspector.setOnlyDisplayRelevantRegionsInCircularView(
-                  e.target.checked,
-                )
-              }
-            />
-          }
-          label="show only regions with data"
-        />
-      </Grid>
-    </Grid>
-  )
-})
 
 export default observer(function SvInspectorView({
   model,
@@ -81,9 +45,12 @@ export default observer(function SvInspectorView({
   } = model
 
   return (
-    <div>
+    <div className={classes.container}>
       <div className={classes.viewsContainer}>
-        <div style={{ width: model.spreadsheetView.width }}>
+        <div
+          style={{ width: model.spreadsheetView.width }}
+          className={classes.container}
+        >
           <SpreadsheetViewReactComponent model={model.spreadsheetView} />
         </div>
